Renames files to fit protocol
